==> src/manual-testing/07_pink_model_selection_01.R <==
library(readr)
library(here)
library(rstanarm)
library(bayesplot)

pink_sr <- read_csv(here("./data/spawner-recruit/clean/pink-sr-for-model.csv"))

# make sure all RE's are factors
pink_sr$brood_year <- as.factor(pink_sr$brood_year)
pink_sr$area <- as.factor(pink_sr$area)
pink_sr$con_unit <- as.factor(pink_sr$con_unit)
pink_sr$river <- as.factor(pink_sr$river)

# bayesian version =============================================================
start_time <- Sys.time()
null_model <- rstanarm::stan_lmer(
  survival ~ spawners:river + (1|brood_year/area) + 
    (-1+con_unit|brood_year) + (1|river),
  data = pink_sr,
  #family = gaussian(link = "identity"),
  #prior = normal(0, 5),
  chains = 3,
  adapt_delta = 0.99,
  control = list(max_treedepth = 15),
  cores = 3,
  iter = 2000
)
null_model$waic <- waic(null_model, cores = 4)
qs::qsave(null_model,
          here("./outputs/model-outputs/TEMP-MANUAL/null-model.qs"))

end_time <- Sys.time() 
end_time - start_time

start_time <- Sys.time()
alt_1 <- rstanarm::stan_lmer(
  survival ~ spawners:river + lice_1  + (1|area_year) + (1|area) + 
    (-1+con_unit|brood_year)  + (1|river),
  data = pink_sr,
  #family = gaussian(link = "identity"),
  #prior = normal(0, 5),
  chains = 3,
  adapt_delta = 0.99,
  control = list(max_treedepth = 15),
  cores = 3,
  iter = 2000
)
alt_1$waic <- waic(alt_1, cores = 4)
qs::qsave(alt_1,
          here("./outputs/model-outputs/TEMP-MANUAL/alt-1.qs"))

end_time <- Sys.time() 
end_time - start_time

start_time <- Sys.time()
alt_2 <- rstanarm::stan_lmer(
  survival ~ spawners:river + lice_2  + (1|area_year) + (1|area) + 
    (-1+con_unit|brood_year)  + (1|river),
  data = pink_sr,
  #family = gaussian(link = "identity"),
  #prior = normal(0, 5),
  chains = 3,
  adapt_delta = 0.99,
  control = list(max_treedepth = 15),
  cores = 3,
  iter = 2000
)
alt_2$waic <- waic(alt_2, cores = 4)
qs::qsave(alt_2,
          here("./outputs/model-outputs/TEMP-MANUAL/alt-2.qs"))

end_time <- Sys.time() 
end_time - start_time
